requirements.txt
discord.py==2.3.2
web3==6.10.0
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
aiohttp==3.8.5
requests==2.31.0
